Move: 쿼리키 파일을 src 폴더 하위에서 constants 폴더 하위로 위치 이동
